Update Formulas.h
Changed `uint32 const nBaseExp = 45` from `unit32` to `float`, per brotalnia

+++ src/game/Formulas.h
@@ -96,7 +96,7 @@ namespace MaNGOS
 
         inline float BaseGain(uint32 ownerLevel, uint32 unitLevel, uint32 mob_level)
         {
-            uint32 const nBaseExp = 45;
+            float const nBaseExp = 45;
             return (ownerLevel * 5 + nBaseExp) * BaseGainLevelFactor(unitLevel, mob_level);
         }
